Moved Idat and MPIClient test impls to separate files
